fix type definition on useResizeObserver
When using `useResizeObserver`, I discovered that the definition of the return type of this hook was misleading. `DOMRect` is initialized to `undefined`, and upon the first run of this hook, the return value can be undefined. In fact, the docs all show that you should check for the existence of `DOMRect` before doing anything with the return value. That should be reflected in the types.